amd/common/amd_pci_util: Extend the routing info to include bus and segment
On multi-domain system there are multiple host bridges. To perform a
proper PCI interrupt routing, the bus number and segment group has to
be included as well in the routing information, because the second and
subsequent host bridges are located at non-zero bus numbers. To
distinguish PCI interrupt routes from these host bridges (and their
subordinary devices) from the first host bridge devices at bus 0, the
devfn in routing info is extended to contain the complete address of
the device, including the bus and segment group.

This fixes the issue where the interrupts were incorrectly assigned on
the whole system, because the routing info was checked for devices on
bus 0.

As AMD FSP is using a byte to represent the devfn of a remapped device,
preserve the old structure for the sake of compatibility with AMD FSP.

TEST=PCI interrupts are assigned properly for all devices under all
host bridges on Gigabyte MZ33-AR1.

Upstream-Status: Backport [CB:89823]

+++ src/southbridge/amd/common/amd_pci_util.h
@@ -11,7 +11,7 @@
 #define PCI_INTR_DATA	0xc01
 
 struct pirq_struct {
-	u8 devfn;
+	u16 devfn;
 	u8 PIN[4];	/* PINA/B/C/D are index 0/1/2/3 */
 };
